+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Move IISW3C formatter and input back from plugin to core.
+- Add SubRip (.srt) files formatter.
 
 ## Fixed